+++ src/main/java/ch/ethz/matsim/av/config/operator/GeneratorConfig.java
@@ -9,11 +9,13 @@ public class GeneratorConfig extends ReflectiveConfigGroup {
 
 	static public final String TYPE = "type";
 	static public final String NUMBER_OF_VEHICLES = "numberOfVehicles";
+	static public final String VEHICLE_TYPE = "vehicleType";
 
 	static public final String DEFAULT_GENERATOR = PopulationDensityGenerator.TYPE;
 	private String type = DEFAULT_GENERATOR;
-	
+
 	private int numberOfVehicles;
+	private String vehicleType;
 
 	public GeneratorConfig() {
 		super(GROUP_NAME, true);
@@ -37,5 +39,15 @@ public class GeneratorConfig extends ReflectiveConfigGroup {
 	@StringSetter(NUMBER_OF_VEHICLES)
 	public void setNumberOfVehicles(int numberOfVehicles) {
 		this.numberOfVehicles = numberOfVehicles;
+	}
+
+	@StringGetter(VEHICLE_TYPE)
+	public String getVehicleType() {
+		return vehicleType;
+	}
+
+	@StringSetter(VEHICLE_TYPE)
+	public void getVehicleType(String vehicleType) {
+		this.vehicleType = vehicleType;
 	}
 }